+++ templates/Pages/view.php
@@ -20,17 +20,17 @@ $this->assign('title', $page->get('title'));
  */
 $class = 'badge badge-warning';
 if (!$page->get('active')) {
-    $this->userbar($this->Html->span(I18N_DRAFT, compact('class')));
+    $this->addToUserbar($this->Html->span(I18N_DRAFT, compact('class')));
 }
 if ($page->get('created')->isFuture()) {
-    $this->userbar($this->Html->span(I18N_SCHEDULED, compact('class')));
+    $this->addToUserbar($this->Html->span(I18N_SCHEDULED, compact('class')));
 }
-$this->userbar($this->Html->link(
+$this->addToUserbar($this->Html->link(
     __d('me_cms', 'Edit page'),
     ['action' => 'edit', $page->get('id'), 'prefix' => ADMIN_PREFIX],
     ['class' => 'nav-link', 'icon' => 'pencil-alt', 'target' => '_blank']
 ));
-$this->userbar($this->Form->postLink(
+$this->addToUserbar($this->Form->postLink(
     __d('me_cms', 'Delete page'),
     ['action' => 'delete', $page->get('id'), 'prefix' => ADMIN_PREFIX],
     ['class' => 'nav-link text-danger', 'icon' => 'trash-alt', 'confirm' => I18N_SURE_TO_DELETE, 'target' => '_blank']