+++ templates/Photos/view.php
@@ -19,14 +19,14 @@ $this->assign('title', $photo->get('filename'));
  * Userbar
  */
 if (!$photo->get('active')) {
-    $this->userbar($this->Html->span(I18N_NOT_PUBLISHED, ['class' => 'badge badge-warning']));
+    $this->addToUserbar($this->Html->span(I18N_NOT_PUBLISHED, ['class' => 'badge badge-warning']));
 }
-$this->userbar($this->Html->link(
+$this->addToUserbar($this->Html->link(
     __d('me_cms', 'Edit photo'),
     ['action' => 'edit', $photo->get('id'), 'prefix' => ADMIN_PREFIX],
     ['class' => 'nav-link', 'icon' => 'pencil-alt', 'target' => '_blank']
 ));
-$this->userbar($this->Form->postLink(
+$this->addToUserbar($this->Form->postLink(
     __d('me_cms', 'Delete photo'),
     ['action' => 'delete', $photo->get('id'), 'prefix' => ADMIN_PREFIX],
     [