+++ templates/PhotosAlbums/view.php
@@ -23,14 +23,14 @@ if (getConfig('default.fancybox')) {
  * Userbar
  */
 if (!$album->get('active')) {
-    $this->userbar($this->Html->span(I18N_NOT_PUBLISHED, ['class' => 'badge badge-warning']));
+    $this->addToUserbar($this->Html->span(I18N_NOT_PUBLISHED, ['class' => 'badge badge-warning']));
 }
-$this->userbar($this->Html->link(
+$this->addToUserbar($this->Html->link(
     __d('me_cms', 'Edit album'),
     ['action' => 'edit', $album->get('id'), 'prefix' => ADMIN_PREFIX],
     ['class' => 'nav-link', 'icon' => 'pencil-alt', 'target' => '_blank']
 ));
-$this->userbar($this->Form->postLink(
+$this->addToUserbar($this->Form->postLink(
     __d('me_cms', 'Delete album'),
     ['action' => 'delete', $album->get('id'), 'prefix' => ADMIN_PREFIX],
     ['class' => 'nav-link text-danger', 'icon' => 'trash-alt', 'confirm' => I18N_SURE_TO_DELETE, 'target' => '_blank']